manual: Review for alt text when images disabled
Added some where the text made no sense at all without the images.

Not much I can do for the specimens, and trying to describe the
example words is probably too much. I could maybe describe the
alphabet in such a way you could draw it yourself, but the
"How to Join" sort of does that already.