Fix messed up [blah]{.ul} rewrites
I identified 2 issues:

* Newlines in the middle of the square brackets
* Multiple underlines in one line

To address newlines, include \n in the match.

To address multiple underlines, use a non-greedy match.

Thus, match on:

    \[((?:[^\]]|\n)+?)\]\{\.ul\}

that is, the minimal not-] or newline characters between square
brackets, followed by the literal {.ul}.
(This is pretty hard to see through all the escaping.)

And rewrite to:

    _$1_